Move compute.TritonError to client.TritonError
The original bug stems from `client.DecodeError()` returning a `client.ClientError{}`
not a `client.TritonError{}`.  Chase relocating `compute.TritonError` into
`client.TritonError` which is the leaf in the package import graph.

Discussed with: @cheapRoc and @stack72
